MasterTokenPool: fix a bug in distributing milestone tokens to subordinate pools.
This bug would distribute fewer tokens than expected to the 2nd, 3rd, ... subordinate pools and give the remainder to the Nth (final) pool. As this bug only occurs when >= 3 subordinate pools are used, it was not detected in testing and would not have occurred in deployment, as only 2 subordinate pools are used. Nevertheless the bug has been corrected and test coverage added for this case.