RewardPools: transfer reward tokens using a `safeTransfer` wrapper around ERC20's `transfer` function.
Some token, e.g. USDT, use a non-standard implementation of the `transfer` function which has no return value. This causes transaction failure for contract code expecting the ERC20 standard interface, even if the return value is not explicitly read.

This change wraps transfers of the reward token with a `safeTransfer` function, allowing correct operation for non-standard tokens. Tokens staked for rewards in the TokenStakeRewardPool do not have this wrapper for deposits but it is assumed they will implement ERC20 exactly.